Bluetooth: Audio: Cast BASS_ACTUAL_SIZE to uint32_t
Adds a cast to the BASS_ACTUAL_SIZE macro so that it can
be printed/logged by `%u` and better match the expected
return value size.

+++ subsys/bluetooth/host/audio/bass_internal.h
@@ -40,7 +40,7 @@
 	((opcode) >= BASS_OP_SCAN_STOP && (opcode) <= BASS_OP_REM_SRC)
 
 #define BASS_ACTUAL_SIZE(x) \
-	(sizeof(x) - (BASS_MAX_METADATA_LEN - (x).metadata_len))
+	((uint32_t)(sizeof(x) - (BASS_MAX_METADATA_LEN - (x).metadata_len)))
 
 struct bass_cp_scan_stop_t {
 	uint8_t opcode;
